Updated openstack/openstack Project: openstack-infra/elastic-recheck  135d7e29e01dcaf2c3161d4af91c9bd402b9d49a
Remove query about yum missing on ubuntu

This query was originaly there to track yum missing on centos6. We no
longer run centos6. The current hits for this query are all from rally
jobs running on ubuntu-trusty. Yum isn't on ubuntu so this is expected.
I think rally must be trying to run yum without checking if it is
present first.

In any case this is just noise on the status page so remove it.

Change-Id: I100b4360b7d6552b938a3e271bd0d53fb726ab37