Updated openstack/openstack Project: openstack/tempest  e4116ec990f5de024475009991d1ec0b48121d68
Do not create network resources in verify-config

Recent changes in devstack/lib/tempest have uncovered this bug
where a neutron router is being created when it isn't needed,
and will fail if a public network hasn't been created.
This is fixed by specifying not to create these network resources
when getting the primary credentials in verify-config.

Change-Id: Ic3857585a74d4e07a9e526e1d96e871fb43154a0
Closes-Bug: #1570097