Updated openstack/openstack Project: openstack/nova  d2cc7d1fe6cd83ec68369905e1801078d81b4ab9
Improve the help text for the network options (1)

The previous commit moved all the options from nova/network/manager.py
to a central location in nova/conf/network.py. This commit adds
additional help text to those options. Because of the large number of
options, the improved help text will be split into several patches; this
is patch 1 of 4.

Blueprint centralize-config-options-newton

Change-Id: I995361157f185a6e03c63e1db858035eb1a31b37