Updated openstack/openstack Project: openstack/neutron-lbaas  f99d74cba51be051a673a5109237295a1bf9c7ac
Set HTTP status code to 409 for LBs in PENDING*

When users attempt to change a LBaaS resource while the load
balancer is in a PENDING-* state they currently get an error
with a 500 HTTP status code.  This patch changes the status code
to be 409 (conflict) which is consistent with other "I'm busy"
errors.

Change-Id: I6fc0c966e72dde956bd481b71a5cea5ba6d10c55
Closes-Bug: #1570596