Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  dee037828944fae2bef3beb71c0fbc6bb3517d25
Use provided mapping object type instead OrderedDict

Better to use the given mapping type from the input data instead of
assuming that the yaml loader has used a particular class type.

This ensures that the code is isolated from any changes made to how the
object type of the input data. Merely needs to know whether it is a
mapping or list type and act accordingly.

Change-Id: If54e7857ca38dc4c9fe7c1f77e8a70f826d39bc8