Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  5d14a621a039a4adf96dfc232b69d0cb574d99be
Allow id's for different templates to follow same naming

It is useful to allow templates that can have different settings for
slightly different jobs while still following the same naming scheme.
Additionally this allows for shorter and more descriptive names to be
used without the confusion of the variable placeholders, and for the
resulting job names from the templates to be altered without needing to
update all references.

Change-Id: Idc3517b44873210a33f988ebff449ea2ed567054