Updated openstack/openstack Project: openstack-infra/zuul  47ef69f941d11d7d3978a4d507b8052b8c52bd6d
Ansible launcher: add zuul_runner module

This runs the commands asynchronously (but waits for their
completion).  This is more robust for long-running commands
because it avoids the built-in ssh timeout.

This adds an ansible module to actually run the remote command
so that we can:
 * process the console log
 * use ansible async (the script module does not support it)
 * control the environment variables of the script being run

It also adds a callback plugin to track the elapsed time so that
we can use the built-in timeout features of async commands.

Note that the module and plugin are GPL licensed.

Change-Id: I19b2b6a5c362bb9d843e7802aefe0eb5df9c5ed7