Updated openstack/openstack Project: openstack/neutron  f2785fc7975bc3e464eef1cf013708df095fa204
Don't update DHCP ports in a transaction

Updating ports is incompatible with the transaction
semantics that happen inside of ML2. See partial-bug for
more details.

This transaction served no purpose wrapped around the port
update because there should only be one port per agent per
network and if that port fails to update, the code to delete
the binding won't be hit anyway.

Change-Id: Ica2c5a19e10ae1773e49537b5ff4ae0eacaeb388
Partial-Bug: #1540844