Updated openstack/openstack Project: openstack/ironic  785c6aca673ff378eb7c69ff76a6d629d32daa78
ipxe: retry on failure

If ipxe fails to boot download kernel/initrd, it will give up and the
boot process will continue with the next boot device, and may end up
with a boot on the old disk image.

If ipxe has failed to download a file and after the pxe/ipxe_timeout
delay, it will now retry to download it instead of giving up.

This goto/retry pattern is common with IPXE. This is an example:
  http://ipxe.org/cmd/autoboot

We can hardly implement complex solutions since we have a very limited
control over the ipxe version in use.

Closes-Bug: #1326656
Change-Id: I0fbb40c711a707ae9fae186e9afbe62b79168e28