Updated openstack/openstack Project: openstack/os-net-config  60475755e3750b2d2e047578a68c8c2327bf50d7
Add warning for no active nics

When there are no active nics, strange errors can happen later
in the process of applying the configuration, and it's often not
obvious what caused them.  Logging a warning should make it
easier to track down such problems.

Note that this should never happen legitimately in a TripleO
environment since we always need to have at least one active nic
to even get configuration to the system.  However, it is a valid
case for someone who might be applying a configuration with local
access to a system, so it should be handled sanely.  It's also
helpful in case of future bugs in the active nic checking.

Change-Id: Iaf6d4b1b215b70d61e0857e093a834702829e1b9
Related-Bug: 1569403