Updated openstack/openstack Project: openstack/cinder  68398b0a3ac10ac8ecd1e4753ab683e8539ed0b3
Use messaging notifications transport instead of default

The usage of oslo_messaging.get_transport is not meant for
notifications; And oslo_messaging.get_notification_transport is
favored for those means. So this change introduces the usage of that
function.

If the settings for the notifications are not set with the
configuration that's under the oslo_messaging_notifications group,
this will fall back to the old settings which are under the DEFAULT
group; just like this used to work.

Change-Id: I1b69e57309df0b6dc675009f42cef34fd501b179