Updated openstack/openstack Project: openstack-infra/project-config  edacb15b7c131141089e016df68150bd4f6d3410
designate-tempest-plugin: Remove server-release-job

We do not need both server-release and publish-to-pypi jobs, it's an
either or. Server projects do not release to PyPI and vice versa.

Having the publish-to-pypi template as well as openstack-server-release-jobs
results in duplicates of some of the jobs running, especially the
tarball and announcement email jobs.

Change-Id: Ic2a82d440577154126218feaaa17cde28ffbf5fa