Updated openstack/openstack Project: openstack-infra/project-config  c87eb42c35a65e0ca9f0a5ed7acb2cd7e58d754f
jjb/puppet: add missing deps on Red Hat systems

To compile some gems, when running Puppet jobs (lint for example), a
package is required: zlib-devel.

It is already installed on Ubuntu, but is not on Red Hat systems.
We did not detect this missing dependency until we added
other-requirements.txt to Puppet modules.

This patch makes sure we install the dependency, so it's
consistent with Ubuntu, and allow us to have other-requirements.txt file
in place for some repo where we don't want bindeps default packages.

Change-Id: I3e57641bd30c6fd2c42881b57180037421fb1272