Updated openstack/openstack Project: openstack-infra/project-config  3b5124bbc3d026a90e0be5a10b3e5084cf34f21a
Neutron repos: Remove extra openstack-server-release-jobs

This fixes a configuration error:

We do not need both server-release and publish-to-pypi jobs, it's an
either or. Server projects do not release to PyPI and vice versa.

Remove openstack-server-release-jobs from repos that do not need it.

Having the publish-to-pypi template as well as
openstack-server-release-jobs results in duplicates of some of the
jobs running, especially the tarball and announcement email jobs.

Change-Id: I4a20febd682617becaf9ba6f5660f4931b8d96f1