Updated openstack/openstack Project: openstack-infra/project-config  7d7ecb9b14cbf942f1b3f164d0e4df5dc9cc3145
Remove publisher site for vendor specific networking-* projects

Documentation of these projects is not meant to be published on
docs.o.o.

Remove openstack-server-publish template and publisher-site.

Also, remove openstack-server-release template for those projects
that also have publish-to-pypi. These are two alternative ways of
publishing - either to pypi or not. Specifying both is wrong.

Note that these two changes go hand in hand to to removal of job-group
from jenkins/jobs/projects.yaml.

Change-Id: Ie154cd0bad33e15c3ce9560d4c67a4c1f33e6f15