Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  88ae2df307e114df98bf12c4c8f3157c8eadec17
Take parameter order from yaml in trigger_parameterized_builds

Support taking the order of the parameters to be passed through to a
triggered parameterized build from the user defined yaml via config
option. Use the existing order by default so as not to break current
usage.

The XML ordering of the parameters matters since when the same parameter
is defined through multiple sources, Jenkins uses the ordering to
determine which will be the one that that should win.

ie if FOO=BAR is set as a build parameter of the current job and a
properties file is generated during build that attempts to set FOO=BAZ;
the XML ordering determines which value, BAZ or BAR, that the triggered
job sees for the parameter FOO.

Change-Id: Ia0dafad8ed3976e7b0b11feb6b3af844bbda9d4a