Updated openstack/openstack Project: openstack/nova  9b2329d10021a7fa5f43b9fd8be3c833dd7212d0
Fail to start nova-api if no APIs were able to be started

This helps to make it clear that either you specified no API services,
or you specified all bad ones, or that you fat-fingered the thing
entirely. If we can't start any, we shouldn't start the process acting
like we did and then wait for you to figure out the hard way that we
aren't actually listening for any requests.

Change-Id: I864486a58cf3fde35647ae3b8d2ad917c809ea54
Related-Bug: #1572495