Updated openstack/openstack Project: openstack-infra/project-config  596b944f16b15b55348675c5357ec8303965aa4e
cloudkitty-dashboard: Remove extra openstack-server-release-jobs

This fixes a configuration error:

We do not need both server-release and publish-to-pypi jobs, it's an
either or. Server projects do not release to PyPI and vice versa.

Remove openstack-server-release-jobs from repos that do not need it.

Having the publish-to-pypi template as well as
openstack-server-release-jobs results in duplicates of some of the
jobs running, especially the tarball and announcement email jobs.

Change-Id: Iea5ffd8860b0541094a7cb1506b3f13d312be5a0