Updated openstack/openstack Project: openstack-infra/project-config  992fae732c40f674a108ca995de18117ddb927d1
Skip gate-rally-dsvm-neutron-neutron in kilo

The job fails because it attempts to install rally using global
requirements from kilo. Rally does not provide us with a stable/kilo
branch to use, while it also requires some dependencies that are not
present in global requirements in kilo (like python-manilaclient).

I don't know of plans to fix the job, so it's probably better to
deprovision it.

Change-Id: Ie7c74ae2917e0da74355e683c2df48cf6ff42a86