Updated openstack/openstack Project: openstack-infra/shade  983f8e1420ae8669e8ba7be2a8858e674f47c99a
Fail if FIP doens't have the requested port_id

When we create a FIP on a port, it should return with the port in
question. Also, it turns out we can wait for the floating IP to
reach ACTIVE status. Finally, our normalize method for fips was
producing inconsistent values depending on create or list - so fix that.

Change-Id: I56b4136a77dd61b6ef1832759b8169dd53aa49da