Updated openstack/openstack Project: openstack/cinder  fb732eeb8db7c67a092c73c8c00b832ce8b0fa67
Microversion of Bootable filter in cinder list

Currently cinder treats all non zero values passed to it
as True for bootable filter in volume GET call.
For ex, if bootable=invalid, it converts it into True being
bootable as boolean value in db table 'volume'.

Need to update filtering functionality in cinder to address
'True' and 'False' values separately and raise exception if
anything than True/False/true/false/0/1 is passed in this filter.

But as this will change the API behavior, microversioning is done
for this fix.

Closes-Bug: #1551535

Change-Id: Idd57014c1504eb4fc9ea0eabe894d2000ea9c364