Updated openstack/openstack Project: openstack/tempest  74d3c1fb4d33d7cd026c4ac61e704a076f1b8340
Add a test for reverting a resize with a deleted flavor

A user should be able to boot a server, resize it and
then reject the resize even though the original flavor
used to boot the instance is deleted. This is because
the old flavor information is stored with the instance
in the nova database, so the original flavor doesn't
actually need to exist anymore.

Depends-On: I5f95021410a309ac07fe9f474cbcd0214d1af208

Change-Id: I356411f96a601f1443d75ac90e42567bef1f8228
Closes-Bug: #1570748