Updated openstack/openstack Project: openstack/neutron  91d95197d892356bd1ab8a96966c11e97d78441b
Revert "ML2: Configure path_mtu to default to 1500 bytes"

This reverts commit 7a4633a9ca213ee03cbcd45027189e0b3e7df2f5.

Revert to using 0 as the default value for path_mtu.

In most situations, underlying MTU does not differ for tunnel backed and
vlan/flat tenant networks, in which case the only configuration expected
from operators is setting global_physnet_mtu to the appropriate MTU
value as reflecting all data paths that tenant traffic may take between
nodes.

But with the non-zero value set for path_mtu, if an operator would like
to raise the global underlying MTU used by neutron to support Jumbo
frames, both global_physnet_mtu and path_mtu need a bump, which is not
ideal.

So switch back to using a zero value for path_mtu, effectively making it
not participating in MTU calculation, unless explicitly overridden.

Left the original release note intact since it reflects the state for
Mitaka.

Added a release note for the change.

Conflicts:
	neutron/plugins/ml2/config.py
	releasenotes/notes/1500-default-mtu-b0d6e4ab193b62a4.yaml

Change-Id: I97c4aa647efc85d7b6b45359e43e2a2ae2514a69