Updated openstack/openstack Project: openstack/python-novaclient  ed645a761868f2a8fbcdf00cd3a35cb1ef00471e
Add a note message for reboot

Currently nova reboot only performs a hard-reboot to
baremetal nodes as ironic doesn't support the soft-reboot
yet. And as absence of this information, end user has to try
to excute this command firstly and then probably realize
the difference between the baremetal nodes and ordinary
VM instance about this command:

 * For VM instance
   nova reboot       : perform a soft reboot by default.
   nova reboot --hard: perform a hard reboot.

 * For baremetal node
   nova reboot       : perform a hard reboot by default.
   nova reboot --hard: perform a hard reboot as well.

For more user-friendly, we'd better add some notes into
the help message explicitly.

Change-Id: I7b3f6da4b5f8a514c6d923f3546470d9de459003
Partial-Bug: #1485416