Updated openstack/openstack Project: openstack/nova  57549b98ac388e0916253cb46e1f5a3dc0961d15
remove the use of import_object_ns

This function dates back to I5eee3389c7719d5f361532b0eddaa249233283a5
when it was introduced to make it easier to seamlessly load drivers
from out of tree. This no long is something we want to do. Also, the
way this works it means that it makes it really hard to debug when an
import of a driver fails (as it's masked out).

This is the only use of the function in all of OpenStack, it should
just go away as an anti-pattern.

Tests were updated to pass, and the compute_driver flag was removed
from tests that passed even when they now had completely bogus values.

Change-Id: I65a537292aa8eda9f3d89e408caa50c445f2050a