Updated openstack/openstack Project: openstack/nova  bc0a84a27ecfb6a1af70b5a1a2e999650fece292
Improve the help text for the linuxnet options (3)

The previous commit moved all the options from network/linuxnet.py to a
central location in nova/conf/network.py. This commit adds additional
help text to those options. Because of the large number of options, the
improved help text will be split into several patches; this is patch 3
of 4.

Blueprint centralize-config-options-newton

Change-Id: I1b5a448b5628f423c4faf58e6e8af978f27e8a8a