Updated openstack/openstack Project: openstack/tripleo-heat-templates  16cb316f691169869956ed58efc41db8b5ebcccb
Use 'deeper' hiera_hash merge behavior for all roles

This change configures the hiera merge behavior to 'deeper' [1],
which is useful to merge values when the same hiera key is found
in multiple datafiles.

The hiera default 'native' only picks the value from the key with
the highest priority in the hierarchy.

1. https://docs.puppetlabs.com/hiera/1/lookup_types.html#deep-merging-in-hiera--120

Change-Id: I88c764d9af510ffbbad9fcaa4b747655e38255c2