Updated openstack/openstack Project: openstack/python-novaclient  c22cbb6fac2230c3111ca60748edcc1591e26c13
[Trivial] change some functions in novaclient/utils.py to public

There are some functions in novaclient/utils.py which have name with
prefix `_`, in Python, this means it is a private function which
should be only used in its own module. However, these functions are
used in other modules such as novaclient/v2/shell.py.

This patch removes the prefix _ for these functions.

Change-Id: I7bc8a76fd390a7dd30eecbb5c7e641b6ccfb40c0