Updated openstack/openstack Project: openstack/python-keystoneclient  641aa0f91e34f19c1cc7880565eb3a3f0da4876e
Fallback if Git repository is absent

Documentation build fails during packaging if Git repository is absent.
We do not package .git directory and that is why it leads to fails during
documentation build.
With this change we are certain that it will not fail.

This change was originally proposed by Davanum Srinivas (dims):
https://review.openstack.org/287448/

Change-Id: I49dce2537ea26c168af9a67d398930042702762c