Updated openstack/openstack Project: openstack/ceilometer  786027eeaca7d890ccf079680121b39451ee3ec5
[Trivial] Remove CEILOMETER_API_LOG_DIR option for devstack

This option is used for ceilometer-api when
CEILOMETER_USE_MOD_WSGI=False, and create a dedicate directory for
ceilometer-api log.

However, there is no such need, because for such case, ceilometer-api
is just a normal process like ceilometer-polling. We should let
oslo.log decide where to store the log, image that user configures
a path and we override in installation process, then he needs to hack
into screenrc to adjust the path, it is not friendly.

Here I propose to put the api log files in same directory as
ceilometer-polling, by default, it will be /opt/stack/logs.

Change-Id: Icca16dbef53929686950a72b40b2544f45f4afdd