Updated openstack/openstack Project: openstack/keystone  37afc8e2dcc483463dccead9a0b433fe89257037
Add test for authentication when project and domain name clash

Now that we store domains as a top level project, it is allowable
to have a project name within than domain to clash with the domain
name itself. This patch adds tests to ensure we can get a project
scoped token to such a project, but not to the project acting as
a domain.

Change-Id: Ib7151633b3a09ab85daba85b0f9853f8c76b1cbf