Updated openstack/openstack Project: openstack-infra/storyboard-webclient  621ed42c72617668fe865361b141e8034bc91814
Don't try to update tasks if they don't have an ID

If the task in the webclient doesn't have an ID, it can't be
updated (obviously). Don't try to update the task in this case.

This patch fixes an error that would arise when setting the status
of a task when creating it on the story detail view.

Change-Id: I0e02ef1a926cc0513d6a8dd85d66e4462e2ce76c