Updated openstack/openstack Project: openstack/zaqar  4dad985c041e37a11efde51e7354feb6438694b0
Default message ttl is needed to change correctly

Currently, when querying queue, Zaqar will return
the _default_message_ttl and _max_messages_post_size,
for _default_message_ttl, it's value is 1209600
if user didn't set it, but it's not correct.
It should be 3600 by default.

APIImpact
Change value of _default message_ttl to 3600

Change-Id: I6b999e56806c6ddcf47190a1fb1953ae373d1ad3
Closes-Bug: #1569663