Updated openstack/openstack Project: openstack/ceilometer  758d3cdc3f3ac5116ccd3e38e522e0432caff1d8
Fix notification listeners usage

Closes-Bug: #1570468

Create the single notification pipeline listener and to aggregate
targets and endpoints in it.

In this implementation the incoming notification msg will be dispatched
in a relevant endpoint by its own filter rule based on publisher id
of a notifier, which sent this message.

Each pipeline notifier has a conforming publisher id based on
a pipeline name in which it to produce messages.

Change-Id: If8c1c6838fd291ef8dbdd4d27c78fe5a36522178