Updated openstack/openstack Project: openstack/oslo-specs  a4280690d117f0fbf70d082c726fad9eb2150ead
versionedobjects: add json schema generation

Versioned objects have a well defined schema that is available
as python code. This schema can also be expressed in a python
independent way for example with JSON Schema. Nova uses versioned
object to define its notification interface. For the consumers of
the notifications it would be super awesome to have the schema
defined in a easy to process and python independent way.

Change-Id: I2073de230f1a74b275087fe7b0c2b4f5953c51b3