Updated openstack/openstack Project: openstack/neutron-vpnaas  36f16b14cac253b8c11a7cf4f0955ca9696eea0e
Support local_id configuration

In some case, we want to specify the local id instead of use the
external ip. For example, the router is behind NAT, and we don't
want to expose this IP to peer. This patch will make local_id
configurable. Most often, local id would be domain name, email
address etc. But if local id is not configured then external ip
will still be used as the leftid.

Note: this change is DocImpact & APIImpact

Change-Id: I2433d72ba22ce410cbfafb7b8bfb16f51eefdfc1
Closes-Bug: #1513353