Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  c4fda8383000439384e7db4034b056a2d6082902
Update Hipchat support to work w/all versions

The HipChat plugin has changed a good bit in version 0.1.9 such that it
is now purely using the publisher hook and it's acquired a few new
configuration items.

This patch ensures used of the existing format will be automatically
redirected to the newer publisher module which adds support for 0.1.9.

Change-Id: I39452e877b20bfc5654d46c90e91dd877a8ebeb0
Story: 2000301
Story: 2000324