Updated openstack/openstack Project: openstack/ceilometer  cd2e0418c7b5f0c52f3c8068e89d1797e365ae5d
devstack: disable workload_partition for compute polling

I7adef87b03129f4f8b38109bf547c7403cc6adad introduces a very good
feature to partition workload for central polling agent, but it
seems not so much value for compute polling agent, which is
natrually workload distributed. There are efforts to reduce the
load of nova-api by adding instance cache in compute polling
agent side, if we increase the compute polling process in the
same node, pollsters may be a bit quicker than before, but things
will get worse for nova-api. And if we want to accelerate the
pollsters, we can search for a better solution instead of increasing
state shared workers.

For most case, I cannot see there is a need to partition workload
on a single node, especially when we running a devstack.

This patch removes the workload_partition setting in devstack
installation for compute polling agent even when there is a
coordinate backend configured. If an user does want this, he can
edit the /etc/ceilometer/ceilometer.conf and restart the compute
polling agent in screen -x, as the usual way.

Change-Id: If4dbcf1d02f5f8bdc189b272fb350a800e23d760