Updated openstack/openstack Project: openstack/manila  e0bd245203654949b7bff318d990ebc9e811bdee
NetApp: DR look up config via host name

In multibackend config, the cDOT driver looks for stanzas with the
share_backend_name as the stanza name (assuming that that's the
hostname, which it isn't).

There needs to be a separation between these since
share_backend_name and stanza names can be different.

Change-Id: Id1c4380c5ac01a0b64645d249a7b61b4c2a62226
Closes-Bug: 1568997