Updated openstack/openstack Project: openstack/governance  587704a743eb67a9faa51435ad7ea21ce7e55cbf
Change kolla to release:cycle-trailing

The openstack/kolla repository has followed the
release:cycle_with_milestones schedule +/- 1-3 days for all
milestones in Liberty and Mitaka except where the PTL was ill
with a root canal that failed causing Mitaka-1 to be tagged
far later then it should have been. In that case, the PTL
added backup release managers (asalkeld inc0 and rhallisey)
so that wouldn't occur in the future.

Now that Doug has proposed cycle-trailing, we can meet that

Depends-On: I783f6f6d55ab33386851910d09bf0e3f0efda404
Change-Id: Ibc6837b46c70c3458f8efdb2e79e7bd3a3b3c189