Updated openstack/openstack Project: openstack/python-swiftclient  6b065dd4a6c82e092382b64bc29f420527a134af
Identify segments uploaded via swiftclient

...using a new "application/swiftclient-segment" content-type.

Segments uploaded by swiftclient are expected to have a many-to-one
relationship to large objects, rather than the more-general many-to-many
relationship that SLO and DLO generally allow. Later, we may use this
information to make more intelligent decisions, such as when to
automatically clean up segments.

Change-Id: Ie56a3aa10065db754ac572cc37d93f2c901aac60