Updated openstack/openstack Project: openstack/tripleo-heat-templates  cab8f750a91238a1e3227d40c81e4c40188a621d
Updates ControlPlaneSubnetCidr to be a string

Heat is unable to execute the following join when creating the
os-net-config definition (OsNetConfigImpl):

	addresses:
	-
		ip_netmask:
		list_join:
		- '/'
		- - {get_param: ControlPlaneIp}
	          - {get_param: ControlPlaneSubnetCidr}

In the default network-environment.yaml environment file the
ControlPlaneSubnetCidr is defaulted as a non-string "24", e.g.

	ControlPlaneSubnetCidr: 24

Resulting in the following error, upon deploymnet:

2016-03-30 22:44:17 [overcloud-Controller-wazxxjc6dq22-2-2ictqfdx2nuo]:
  CREATE_FAILED  Resource CREATE failed:
    resources.NetworkConfig:
      Property error:
        resources.OsNetConfigImpl.properties.config:
          Items to join must be strings not 24

This patch simply updates the default value to '24', to match other
templates. This allows a deployment to succeed as expected.

Change-Id: I5f75b02dba5c35a9856c6ceff9cf6c24a0cb56d8